Pull out common loading code to utils
* Update individual loaders to use the utils function
  . Completed med hist, motor, and sub char loaders
* Update test text and specifics to match the harmonized version
